requirements.txt
aggdraw==1.3.16
attrs==23.1.0
docopt==0.6.2
imageio==2.31.1
lazy_loader==0.3
networkx==3.1
numpy==1.25.1
packaging==23.1
Pillow==10.0.0
psd-tools==1.9.28
PyWavelets==1.4.1
scikit-image==0.21.0
scipy==1.11.1
tifffile==2023.7.18
